Move long CI DOCI tests
